Gemma 3: Force cast bfloat16 to np.float64

+++ semantic_uncertainty/uncertainty/uncertainty_measures/p_ik.py
@@ -17,7 +17,7 @@ def get_p_ik(train_embeddings, is_false, eval_embeddings=None, eval_is_false=Non
     # Convert the list of tensors to a 2D tensor.
     train_embeddings_tensor = torch.cat(train_embeddings, dim=0)  # pylint: disable=no-member
     # Convert the tensor to a numpy array.
-    embeddings_array = train_embeddings_tensor.cpu().numpy()
+    embeddings_array = train_embeddings_tensor.cpu().float().numpy() # Force typecast half-precision bfloat to float before converting to numpy array.
 
     # Split the data into training and test sets.
     X_train, X_test, y_train, y_test = train_test_split(  # pylint: disable=invalid-name
@@ -28,7 +28,7 @@ def get_p_ik(train_embeddings, is_false, eval_embeddings=None, eval_is_false=Non
     model.fit(X_train, y_train)
 
     # Predict deterministically and probabilistically and compute accuracy and auroc for all splits.
-    X_eval = torch.cat(eval_embeddings, dim=0).cpu().numpy()  # pylint: disable=no-member,invalid-name
+    X_eval = torch.cat(eval_embeddings, dim=0).cpu().float().numpy()  # pylint: disable=no-member,invalid-name # Force typecast half-precision bfloat to float before converting to numpy array.
     y_eval = eval_is_false
 
     Xs = [X_train, X_test, X_eval]  # pylint: disable=invalid-name
